# Replace HTML5small with Htmlcompressor