--- a/Gemfile
+++ b/Gemfile
@@ -16,7 +16,7 @@ gem 'sass', '~> 3.2.13'
 gem 'susy'        # for nice grids
 gem 'systemu'     # for invoking rsync etc
 gem 'yard'        # for loading filter and helper docs
-gem 'html5small'
+gem 'htmlcompressor'
 gem 'nanoc-typohero'
 gem 'w3c_validators'
 gem 'ffi-aspell'