--- a/lib/filters/htmlcompressor.rb
+++ b/lib/filters/htmlcompressor.rb
@@ -1,0 +1,10 @@
+require 'htmlcompressor'
+
+Class.new(Nanoc::Filter) do
+  identifier :htmlcompressor
+
+  def run(content, params = {})
+    compressor = HtmlCompressor::Compressor.new
+    compressor.compress(content)
+  end
+end