--- a/lib/filters_.rb
+++ b/lib/filters_.rb
@@ -1,2 +1,1 @@
-require 'html5small/nanoc'
 require 'nanoc-typohero'