This cuts down compilation speed by about half. [Htmlcompressor](https://github.com/paolochiodi/htmlcompressor) is significantly faster than [HTML5small](https://github.com/RubenVerborgh/HTML5small). It does generate slightly larger output, though.

Before:

```
Site compiled in 15.28s.
Site compiled in 14.05s.
Site compiled in 15.88s.
```

After:

```
Site compiled in 7.49s.
Site compiled in 7.21s.
Site compiled in 7.15s.
```