@ddfreyne 2016-02-12
CC @RubenVerborgh — I feel that the slightly better output of HTML5small doesn’t justify the increased compilation speed.