@RubenVerborgh 2016-02-12
Any stats on the size difference? Might migrate too and abandon HTML5small. From what I see, the internal workings are similar. (My involvement in HTML5small was just a set of minor fixes, so I don't know too much about its internals.)

BTW I have a condition in my rule file so that compression is only triggered on the final server, so the time does not impact me as much.