@ddfreyne 2016-02-12
From what I can tell, it concerns mostly the optional end tags (`</li>` and `</p>` come to mind) which Htmlcompressor does not filter out.