@RubenVerborgh 2016-02-12
Alright, then we don't need HTML5small anymore. I'll deprecate the project.